Move InstanceSerialConsoleRequest/Data to separate copies in nexus::external_api and sled_agent params mods
