+++ common/src/api/internal/nexus.rs
@@ -103,31 +103,3 @@ pub struct UpdateArtifact {
 pub enum UpdateArtifactKind {
     Zone,
 }
-
-/// Forwarded to a sled agent to request the contents of an Instance's serial console.
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema, PartialEq)]
-pub struct InstanceSerialConsoleRequest {
-    /// Character index in the serial buffer from which to read, counting the bytes output since
-    /// instance start. If this is not provided, `most_recent` must be provided, and if this *is*
-    /// provided, `most_recent` must *not* be provided.
-    pub from_start: Option<u64>,
-    /// Character index in the serial buffer from which to read, counting *backward* from the most
-    /// recently buffered data retrieved from the instance. (See note on `from_start` about mutual
-    /// exclusivity)
-    pub most_recent: Option<u64>,
-    /// Maximum number of bytes of buffered serial console contents to return. If the requested
-    /// range runs to the end of the available buffer, the data returned will be shorter than
-    /// `max_bytes`.
-    pub max_bytes: Option<u64>,
-}
-
-/// Contents of an Instance's serial console buffer.
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-pub struct InstanceSerialConsoleData {
-    /// The bytes starting from the requested offset up to either the end of the buffer or the
-    /// request's `max_bytes`. Provided as a u8 array rather than a string, as it may not be UTF-8.
-    pub data: Vec<u8>,
-    /// The absolute offset since boot (suitable for use as `byte_offset` in a subsequent request)
-    /// of the last byte returned in `data`.
-    pub last_byte_offset: u64,
-}
